// apps/backend/src/types/constants.ts
export enum Services {
	Prisma = "PRISMA_SERVICE",
	Auth = "AUTH_SERVICE",
	User = "USER_SERVICE",
	Transaction = "TRANSACTION_SERVICE",
}

export enum Repositories {
	User = "USER_REPOSITORY",
	Transaction = "TRANSACTION_REPOSITORY",
}

export enum Routes {
	Auth = "auth",
	User = "users",
	Transaction = "transactions",
}